Unfinished work on adding option to set the max depth of tree loading
adding paging control buttons

extensions class renamed to split new extensions later
