+++ JsonTranslationEditor/MainWindow.xaml
@@ -89,7 +89,7 @@
             </ScrollViewer>
         </Grid>
 
-        <GridSplitter Grid.Column="1" Grid.Row="1" HorizontalAlignment="Stretch" Width="5" />
+        <GridSplitter Background="Bisque" Grid.Column="1" Grid.Row="1" Grid.RowSpan="2" HorizontalAlignment="Stretch" Width="5" />
         <TextBox Grid.Column="2" x:Name="SearchFilterTextbox" />
         <ScrollViewer  Grid.Column="2" Grid.Row="1" x:Name="ContentScroller">
             <StackPanel Background="#d4d8d9" x:Name="PagingContainer">
@@ -147,13 +147,24 @@
 
         </ScrollViewer>
 
-        <Grid Grid.Column="2" Grid.Row="3">
+        <Grid Grid.Column="2" Grid.Row="2" Height="20" VerticalAlignment="Bottom" Margin="10,0,0,0">
             <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="100"></ColumnDefinition>
                 <ColumnDefinition Width="*"></ColumnDefinition>
+
+                <ColumnDefinition Width="150"></ColumnDefinition>
+                <ColumnDefinition Width="100"></ColumnDefinition>
             </Grid.ColumnDefinitions>
-            <Button Visibility="Hidden" Grid.Column="0" x:Name="partialPagingButton" Click="ShowAll">Load All</Button>
-            <TextBlock Grid.Column="1"  x:Name="pagingMessage"></TextBlock>
+
+            <TextBlock Grid.Column="0"  x:Name="pagingMessage"></TextBlock>
+            <UniformGrid x:Name="pagingButtons" Visibility="Hidden" Rows="1" Grid.Column="1">
+                <Button Click="FirstPage">&lt;&lt;</Button>
+                <Button Click="PreviousPage">&lt;</Button>
+                <Button Click="NextPage">&gt;</Button>
+                <Button Click="LastPage">&gt;&gt;</Button>
+            </UniformGrid>
+            <Button Visibility="Hidden" Grid.Column="2" x:Name="partialPagingButton" Click="ShowAll">Load</Button>
+            
+
         </Grid>
       
     </Grid>
