+++ JsonTranslationEditor/SummaryInfo.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using JsonTranslationEditor.Extensions;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
